added option in model.py to compute dkappa_dz numerically
dkappa_dz will be computed numerically if kappa is not constant
but dkappa_dz is not provided explicity.
Notice that this only works with numpy version 1.13 and later,
as numpy.gradient can't handle non-even grid in earlier versions.
It would be good to incorporate a more meaningful error message if
use is attempted with earlier numpy version