Solution and grid now part of model instance
Instead of model.solve() returning a dictionary containing
the solution, the solution now is part of the model instance,
and model.solve() simply updates it. The solution can either
be provided on the mesh chosen by the solver, or on a user-defined
grid. I believe this infrastructure will simplify and be more consistent
with some future developments, such as the option to provide kappa
and psi_so on a user-defined grid.