changed initial value of Psi_SO.Psi to None
I don't think we want Psi to have a usable value before solve has been called. A vector of zeros would generally be inconsistent with the prescribed buoyancy profiles, and a user may not realize that the streamfunction has not actually been calculated yet and therefore should not be used. (Alternatively we could call teh solver as part of the initialization)